Create Profile
I am student in filed of the master of control engineering in the Shiraz university. 
My main research and work is about of robotics and control of complex systems. I focus on intelligent algorithms
for control and predict of systems with high complexity such as medical system, robots and etc.
I am currently working on my thesis and the subject is about of control autonomous mobile robots. 
On the other hand, I am mechatronic department manager of Arioobarzan company in Iran that is working 
on the manufacturing of intelligent hardware and software device. Also i am member of robotic and control lab in the Shiraz university.